Reorganize the files in the repo a bit
