+++ web-app/src/components/note/CardNoteCanvasView.ts
@@ -1,11 +1,11 @@
-import { registerElementView } from "../../elementViewLibrary";
+import { registerElementView } from "../elementViewLibrary";
 import { CardNote } from "../../model/CardNote";
 import { ImageNote } from "../../model/ImageNote";
 import { TextNote } from "../../model/TextNote";
 import { calculateTextLayout } from "./util";
 import { NoteCanvasView } from "./NoteCanvasView";
 import { textRect } from "./util";
-import { DEFAULT_FONT_STRING } from "../../constants";
+import { DEFAULT_FONT_STRING } from "../../core/constants";
