+++ web-app/src/components/note/InternalLinkCanvasView.ts
@@ -1,9 +1,9 @@
-import { registerElementView } from "../../elementViewLibrary";
+import { registerElementView } from "../elementViewLibrary";
 import { InternalLinkNote } from "../../model/InternalLinkNote";
 import { calculateTextLayout } from "./util";
 import { NoteCanvasView, BorderType } from "./NoteCanvasView"
 import { textRect } from "./util";
-import { DEFAULT_FONT_STRING } from "../../constants";
+import { DEFAULT_FONT_STRING } from "../../core/constants";
 
 
 export class InternalLinkNoteCanvasView extends NoteCanvasView {
