+++ web-app/src/components/note/ScriptNoteCanvasView.ts
@@ -1,11 +1,11 @@
-import { registerElementView } from "../../elementViewLibrary";
+import { registerElementView } from "../elementViewLibrary";
 import { ScriptNote } from "../../model/ScriptNote";
 import { color_to_css_rgba_string } from "../../util";
 import { calculateTextLayout } from "./util";
 import { Point2D } from "../../util/Point2D";
 import { NoteCanvasView } from "./NoteCanvasView";
 import { textRect } from "./util";
-import { DEFAULT_FONT_STRING } from "../../constants";
+import { DEFAULT_FONT_STRING } from "../../core/constants";
 
 const TRIANGLE_BASE = 10;
 const TRIANGLE_SPACING = 3;
