+++ web-app/src/components/page/PageView.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect, useCallback, useRef } from 'react';
 import styled from 'styled-components';
 import { observer } from 'mobx-react-lite';
 
-import { DEFAULT_EYE_HEIGHT, MAX_HEIGHT_SCALE, MIN_HEIGHT_SCALE } from '../../constants';
+import { DEFAULT_EYE_HEIGHT, MAX_HEIGHT_SCALE, MIN_HEIGHT_SCALE } from '../../core/constants';
 import { Point2D } from '../../util/Point2D';
 import { pageActions } from '../../actions/page';
 import { TourComponent } from '../Tour';
@@ -14,10 +14,10 @@ import React from 'react';
 import paper from 'paper';
 import { CanvasPageRenderer } from './DirectRenderer';
 import { ElementViewState } from './ElementViewState';
-import { pamet } from '../../facade';
+import { pamet } from '../../core/facade';
 import { NavigationDevice, NavigationDeviceAutoSwitcher } from './NavigationDeviceAutoSwitcher';
 // import { CanvasReactComponent } from './ReactRenderingComponent';
-import { commands } from '../../commands';
+import { commands } from '../../core/commands';
 import EditComponent from '../note/EditComponent';
 import { NoteViewState } from '../note/NoteViewState';
 import { Size } from '../../util/Size';
