+++ web-app/src/core/constants.ts
@@ -1,4 +1,4 @@
-import { ColorData } from "./util";
+import { ColorData } from "../util";
 
 export const DEFAULT_TEXT_COLOR: ColorData = [0, 0, 1, 1];
 export const DEFAULT_BACKGROUND_COLOR: ColorData = [0, 0, 1, 0.1];
