+++ web-app/src/model/TextNote.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_BACKGROUND_COLOR, DEFAULT_NOTE_HEIGHT, DEFAULT_NOTE_WIDTH, DEFAULT_TEXT_COLOR } from "../constants";
+import { DEFAULT_BACKGROUND_COLOR, DEFAULT_NOTE_HEIGHT, DEFAULT_NOTE_WIDTH, DEFAULT_TEXT_COLOR } from "../core/constants";
 import { entityType, getEntityId } from "pyfusion/libs/Entity";
 import { currentTime, timestamp } from "pyfusion/util";
 import { elementId } from "./Element";
