--- web-app/src/services/routingService.ts
@@ -1,11 +1,0 @@
-//   // Set the page initially from the URL and subscribe to URL changes
-//   useEffect(() => {
-//     updatePageFromUrl();
-//     window.addEventListener("popstate", updatePageFromUrl);
-
-//     return () => {
-//       window.removeEventListener("popstate", updatePageFromUrl);
-//     };
-//   }, [updatePageFromUrl]);
-
-export {}
