--- web-app/src/storage/pametInMemoryRepo.ts
@@ -1,8 +1,0 @@
-// import { InMemoryStore } from "pyfusion/storage/InMemoryStore";
-// import { PametStore } from "./base";
-
-// export class PametInMemoryStore extends PametStore {
-//     constructor() {
-//         super(new InMemoryStore());
-//     }
-// }
